requirements.txt
gevent
Flask
pocsuite
pymongo
python-nmap
apscheduler
multiprocessing
dnspython
ipaddr
bs4
requests
pexpect
MySQL-python